+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.1.2 (2016-10-01)
+
+* Main: No more default options
+* Main: `GetDefaultOptions()` is now public
+* CLI: `-set` is no longer enabled by default
+* CLI: You can only use `-set` or `-all` at any given time
+* CLI: Show error when no results & print error instead of panic
+
 ## v0.1.1 (2016-09-28)
 
 * Fix CLI